code/requirements.txt
pandas==2.0.2
azure-storage-blob==12.16.0